requirements-bq.txt
pytest
pytest-cov
google-cloud-bigquery
-e .